improv: move terminal security to separate package
