+++ core/package.json
@@ -62,6 +62,7 @@
     "@continuedev/fetch": "file:../packages/fetch",
     "@continuedev/llm-info": "file:../packages/llm-info",
     "@continuedev/openai-adapters": "file:../packages/openai-adapters",
+    "@continuedev/terminal-security": "file:../packages/terminal-security",
     "@modelcontextprotocol/sdk": "^1.12.0",
     "@mozilla/readability": "^0.5.0",
     "@octokit/rest": "^20.1.1",
