+++ extensions/cli/src/ui/hooks/useChat.stream.helpers.ts
@@ -1,4 +1,5 @@
 import type { ChatHistoryItem, ToolCallState, ToolStatus } from "core/index.js";
+
 import { getAllBuiltinTools } from "src/tools/index.js";
 
 interface CreateStreamCallbacksOptions {
