+++ packages/terminal-security/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@continuedev/terminal-security",
+  "version": "1.0.0",
+  "description": "Terminal command security evaluation for Continue",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "type": "module",
+  "scripts": {
+    "test": "vitest run",
+    "test:watch": "vitest watch",
+    "build": "tsc",
+    "clean": "rm -rf dist"
+  },
+  "author": "Continue Dev, Inc",
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "@types/node": "^20.11.19",
+    "@types/shell-quote": "^1.7.5",
+    "typescript": "^5.5.2",
+    "vitest": "^1.6.0"
+  },
+  "dependencies": {
+    "shell-quote": "^1.8.1"
+  }
+}
