+++ packages/terminal-security/src/index.ts
@@ -1,0 +1,2 @@
+export { ToolPolicy } from "./types.js";
+export { evaluateTerminalCommandSecurity } from "./evaluateTerminalCommandSecurity.js";
