+++ packages/terminal-security/src/types.ts
@@ -1,0 +1,7 @@
+/**
+ * Policy for tool execution
+ */
+export type ToolPolicy =
+  | "allowedWithPermission"
+  | "allowedWithoutPermission"
+  | "disabled";
